chore: melhorando a organização do código para maior mantenabilidade
